move category form to new component
